cms-admin: Colocate action log dialog fragment with ActionLogDialog
The ActionLogRow fragment lived in a standalone actionLogsQuery.ts,
didn't belong to any component and listed all fields flat. Replace it
with an ActionLogDialog fragment in ActionLogDialog.gql.ts that is
composed of the fragments of the components the dialog renders
(ActionLogVersionGrid, ActionLogShowVersion, ActionLogCompare), so
every fetched field is owned by a component. This also drops the
entityId and scope fields no component consumed.

Move ActionLogQueryName into ActionLogDialog.tsx, whose props it
describes, and stop exporting the fragment from the package's public
API since apps only interact with ActionLogButton/ActionLogDialog via
queryName.
