+++ packages/admin/cms-admin/src/actionLog/actionLog/actionLogDialog/ActionLogDialog.tsx
@@ -6,14 +6,26 @@ import { FormattedMessage, useIntl } from "react-intl";
 import { type ContentScope, useContentScope } from "../../../contentScope/Provider";
 import { ActionLogCompare } from "../../components/actionLogCompare/ActionLogCompare";
 import { ActionLogShowVersion } from "../../components/actionLogShowVersion/ActionLogShowVersion";
-import { type ActionLogQueryName, buildActionLogsQuery } from "../actionLogsQuery";
-import type { GQLActionLogRowFragment } from "../actionLogsQuery.generated";
 import { ActionLogVersionGrid } from "../actionLogVersionGrid/ActionLogVersionGrid";
+import { buildActionLogsQuery } from "./ActionLogDialog.gql";
+import type { GQLActionLogDialogFragment } from "./ActionLogDialog.gql.generated";
 
 type ActionLogDialogView =
     | { type: "grid" }
-    | { type: "showVersion"; row: GQLActionLogRowFragment }
-    | { type: "compareVersions"; before: GQLActionLogRowFragment; after: GQLActionLogRowFragment };
+    | { type: "showVersion"; row: GQLActionLogDialogFragment }
+    | { type: "compareVersions"; before: GQLActionLogDialogFragment; after: GQLActionLogDialogFragment };
+
+/**
+ * Keys of `TQuery` whose value is a `PaginatedActionLogs` — i.e. the top-level
+ * `<entity>ActionLogs` queries generated by `ActionLogsModule.forFeature()`.
+ * Falls back to `string` when no generic is supplied.
+ */
+export type ActionLogQueryName<TQuery> = string extends keyof TQuery
+    ? string
+    : {
+          [K in keyof TQuery]: NonNullable<TQuery[K]> extends { __typename?: "PaginatedActionLogs" } ? K : never;
+      }[keyof TQuery] &
+          string;
 
 export type ActionLogDialogProps<TQuery> = {
     /**
@@ -32,7 +44,7 @@ export type ActionLogDialogProps<TQuery> = {
 };
 
 type ActionLogsQueryResult = {
-    [key: string]: { nodes: GQLActionLogRowFragment[]; totalCount: number };
+    [key: string]: { nodes: GQLActionLogDialogFragment[]; totalCount: number };
 };
 
 export function ActionLogDialog<TQuery = Record<string, unknown>>({ queryName, entityId, name, open, onClose }: ActionLogDialogProps<TQuery>) {
