--- packages/admin/cms-admin/src/actionLog/actionLog/actionLogsQuery.ts
@@ -1,54 +1,0 @@
-import { gql } from "@apollo/client";
-import { parse } from "graphql";
-
-import { actionLogCompareFragment } from "../components/actionLogCompare/ActionLogCompare";
-
-export const actionLogRowFragment = gql`
-    fragment ActionLogRow on ActionLog {
-        id
-        entityName
-        entityId
-        version
-        type
-        createdAt
-        scope
-        snapshot
-        previousVersion {
-            id
-        }
-        user {
-            id
-            name
-        }
-        ...ActionLogCompare
-    }
-    ${actionLogCompareFragment}
-`;
-
-/**
- * Keys of `TQuery` whose value is a `PaginatedActionLogs` — i.e. the top-level
- * `<entity>ActionLogs` queries generated by `ActionLogsModule.forFeature()`.
- * Falls back to `string` when no generic is supplied.
- */
-export type ActionLogQueryName<TQuery> = string extends keyof TQuery
-    ? string
-    : {
-          [K in keyof TQuery]: NonNullable<TQuery[K]> extends { __typename?: "PaginatedActionLogs" } ? K : never;
-      }[keyof TQuery] &
-          string;
-
-export function buildActionLogsQuery(queryName: string) {
-    const operationName = queryName[0].toUpperCase() + queryName.slice(1);
-    const fragmentBody = actionLogRowFragment.loc?.source.body ?? "";
-    return parse(`
-        query ${operationName}($scope: JSONObject!, $offset: Int!, $limit: Int!, $filter: ActionLogFilter, $sort: [ActionLogSort!]) {
-            ${queryName}(scope: $scope, offset: $offset, limit: $limit, filter: $filter, sort: $sort) {
-                nodes {
-                    ...ActionLogRow
-                }
-                totalCount
-            }
-        }
-        ${fragmentBody}
-    `);
-}
