+++ packages/admin/cms-admin/src/index.ts
@@ -1,6 +1,5 @@
 export { ActionLogButton } from "./actionLog/actionLog/actionLogButton/ActionLogButton";
 export { ActionLogDialog } from "./actionLog/actionLog/actionLogDialog/ActionLogDialog";
-export { actionLogRowFragment } from "./actionLog/actionLog/actionLogsQuery";
 export { ActionLogVersionGrid, actionLogVersionGridFragment } from "./actionLog/actionLog/actionLogVersionGrid/ActionLogVersionGrid";
 export { ActionLogCompare, actionLogCompareFragment } from "./actionLog/components/actionLogCompare/ActionLogCompare";
 export { ActionLogHeader, type ActionLogHeaderProps } from "./actionLog/components/actionLogHeader/ActionLogHeader";
